Apply ZATCA Feedback on Signed Properties
ZATCA’s latest update is that SignedProperties should not be canonicalized anymore. You should use a specific version of the XML element with specific whitespace and add a single attribute only when hashing SignedProperties but remove it in all other instances.

We now use a HEREDOC just for generating this version of the XML for hashing because it does not have consistent indentation.

See:
https://zatca1.discourse.group/t/what-do-signed-properties-look-like-when-hashing
https://web.archive.org/web/20230925182417/https://zatca1.discourse.group/t/what-do-signed-properties-look-like-when-hashing/717