Always Set Qualifying Properties Indentation Matching ZATCA’s
After generating the XML, we modify the QualifyingProperties block’s indentation to match ZATCA’s. ZATCA’s server performs hashing serverside on the SignedProperties as-is without transforming them to their required indentation format.

This forces us to have to make the SignedProperties have the same identation as in their files.